Moved specific JS and handled all errors
