+++ app/javascript/packs/application.js
@@ -1,4 +1,1 @@
 // import "bootstrap";
-import { subnav } from '../components/subnav';
-
-subnav();
