+++ app/views/campaigns/components/_campaign_preview.html.erb
