+++ app/views/campaigns/index.html.erb
@@ -60,33 +60,12 @@
 
 <!-- ========== Page Specific Javascript ========== -->
 
-<%= javascript_include_tag 'vendor/svg-injector/dist/svg-injector.min' %>
-<%= javascript_include_tag 'components/hs.svg-injector' %>
-
-<%= javascript_include_tag 'components/hs.header' %>
-<%= javascript_include_tag 'components/hs.unfold' %>
-
 <%= javascript_include_tag 'vendor/fancybox/jquery.fancybox.min' %>
 <%= javascript_include_tag 'components/hs.fancybox' %>
 
+<%= javascript_pack_tag 'subnav' %>
+
 <script>
-  $(window).on('load', function () {
-    // initialization of svg injector module
-    $.HSCore.components.HSSVGIngector.init('.js-svg-injector');
-  });
-
-  $(document).on('ready', function () {
-	  // initialization of header
-	  $.HSCore.components.HSHeader.init($('#header'));
-
-	  // initialization of unfold component
-	  $.HSCore.components.HSUnfold.init($('[data-unfold-target]'), {
-	    afterOpen: function () {
-	      $(this).find('input[type="search"]').focus();
-	    }
-	  });
-	});
-
   $(document).on('ready', function () {
     // initialization of popups
     $.HSCore.components.HSFancyBox.init('.js-fancybox');
