+++ app/views/investments/new.html.erb
@@ -133,6 +133,8 @@
 <!-- End Shopping Cart Modal Window -->
 </main>
 
+<%= render 'shared/sidebar' %>
+
 <!-- ========== Page Specific Javascript ========== -->
 
 <%= javascript_include_tag 'vendor/ion-rangeslider/js/ion.rangeSlider.min' %>
