+++ app/views/pages/home.html.erb
@@ -1,6 +1,7 @@
 <!-- ========== HEADER ================ -->
 <%= render 'pages/components/home_header' %>
 <!-- ========== END HEADER ============ -->
+
 <!-- ========== MAIN CONTENT ========== -->
 <main id="content" role="main">
  <%= render 'pages/components/home_hero_section' %>
@@ -16,16 +17,21 @@
 <%= render 'pages/components/home_how_section' %>
 </main>
 <!-- ========== END MAIN CONTENT ==== -->
+
 <!-- ========== FOOTER ============== -->
 <%= render 'pages/components/home_footer' %>
 <!-- ========== END FOOTER ========== -->
+
 <!-- ========== Page Specific Javascript ========== -->
+
 <%= javascript_include_tag 'vendor/fancybox/jquery.fancybox.min' %>
 <%= javascript_include_tag 'components/hs.fancybox' %>
+
 <script>
- $(document).on('ready', function () {
-   // initialization of popups
-   $.HSCore.components.HSFancyBox.init('.js-fancybox');
- });
+  $(document).on('ready', function () {
+    // initialization of popups
+    $.HSCore.components.HSFancyBox.init('.js-fancybox');
+  });
 </script>
+
 <!-- ========== End Page Specific Javascript ====== -->
