+++ app/views/pages/user_profile.html.erb
@@ -42,34 +42,4 @@
 <!-- End Breadcrumb Section -->
 <!-- ====== End Sub Navigation ====== -->
 
-
 <%= render 'shared/sidebar' %>
-
-<!-- ========== Page Specific Javascript ========== -->
-
-<%= javascript_include_tag 'vendor/svg-injector/dist/svg-injector.min' %>
-<%= javascript_include_tag 'components/hs.svg-injector' %>
-
-<%= javascript_include_tag 'components/hs.header' %>
-<%= javascript_include_tag 'components/hs.unfold' %>
-
-<script>
-  $(window).on('load', function () {
-    // initialization of svg injector module
-    $.HSCore.components.HSSVGIngector.init('.js-svg-injector');
-  });
-
-  $(document).on('ready', function () {
-	  // initialization of header
-	  $.HSCore.components.HSHeader.init($('#header'));
-
-	  // initialization of unfold component
-	  $.HSCore.components.HSUnfold.init($('[data-unfold-target]'), {
-	    afterOpen: function () {
-	      $(this).find('input[type="search"]').focus();
-	    }
-	  });
-	});
-</script>
-
-<!-- ========== End Page Specific Javascript ====== -->
